+++ qcfractal/interface/models/__init__.py
@@ -4,8 +4,8 @@ Either pull in QCEl models or local models
 
 from . import rest_models
 from .model_builder import build_procedure
-from .gridoptimization import GridOptimization, GridOptimizationInput
+from .gridoptimization import GridOptimizationRecord, GridOptimizationInput
 from .torsiondrive import TorsionDrive, TorsionDriveInput
-from .records import OptimizationRecord
+from .records import ResultRecord, OptimizationRecord
 from .common_models import Molecule, KeywordSet, ObjectId, QCSpecification
 from .model_utils import hash_dictionary, json_encoders, prepare_basis