+++ qcfractal/interface/models/tests/test_records.py
@@ -1,0 +1,8 @@
+import pytest
+from pydantic import BaseModel, ValidationError
+
+from ..records import OptimizationRecord
+import qcelemental as qcel
+
+oid1 = "000000000000000000000000"
+oid2 = "000000000000000000000001"